refactor: move field context to src/contexts/
- Move src/lib/field-context.ts to src/contexts/field.ts
- Add src/contexts/index.ts for centralized context exports
- Update all component imports
