+++ src/components/lv1/Textarea/Textarea.tsx
@@ -1,5 +1,5 @@
 import { forwardRef, type TextareaHTMLAttributes, useId } from 'react'
-import { useFieldContext } from '../../../lib/field-context'
+import { useFieldContext } from '../../../contexts/field'
 import { cn } from '../../../lib/utils'
 import { type TextareaVariants, textareaVariants } from '../../../variants/textarea'
